icons: compact paletted blob (ABI v22) — ~halve the per-game flash cost
Launcher icons were a raw uint16 mote_game_icon_data[3600] = 7,200 bytes in flash.
From ABI v22 they're a compact paletted blob (sdk/mote_icon.h): [bpp][npal][palette]
[packed indices], adaptive 1/2/4/8 bpp. Lossless for pixel-art icons (<=16 colours ->
~1.8 KB); busier icons quantise to a 256-colour RGB332 palette (~3.7-4.1 KB). Decode is
a palette lookup — no decompressor in firmware.

- sdk/mote_icon.h: shared encode (baker) + decode (launcher).
- Bakers emit the blob: tools/mote bake_icon (Python) and studio mc_bake bake_icon (C).
- The launcher decodes the selected game's icon into one 7,200-byte scratch and blits
  it; raw (<=v21) icons still blit straight from flash. Format is implied by
  abi_version, so MoteModuleHeader is unchanged and old stored games still show.
- Device (store_icon) + host (host_fill) route icon vs icon_blob by the module's ABI.
- Rebaked all example icons (the 30 icon.h are now blobs).

Needs a firmware reflash to take effect on device (v22).

+++ sdk/mote_module.h
@@ -36,10 +36,11 @@ typedef struct {
     uint32_t data_end;         /* RAM dest end */
     uint32_t bss_start;        /* RAM zero start */
     uint32_t bss_end;          /* RAM zero end */
-    /* Per-game launcher icon: the linked (VADDR-relative) address of a 60x60
-     * RGB565 `mote_game_icon_data[]`, or 0 if the game ships no icon. The
-     * launcher reads it straight from flash — image_offset + (icon_vaddr -
-     * MOTE_MODULE_VADDR) — without loading the game. */
+    /* Per-game launcher icon: the linked (VADDR-relative) address of
+     * `mote_game_icon_data[]`, or 0 if the game ships no icon. The launcher reads
+     * it straight from flash — image_offset + (icon_vaddr - MOTE_MODULE_VADDR).
+     * Format follows abi_version: a raw 60x60 RGB565 array up to v21, a compact
+     * paletted blob (sdk/mote_icon.h) from v22 — no struct change either way. */
     uint32_t icon_vaddr;
 } MoteModuleHeader;
 
@@ -56,7 +57,7 @@ typedef struct {
 #define MOTE_MODULE_HEADER()                                                       \
     extern char __mote_data_load[], __mote_data_start[], __mote_data_end[],      \
                 __mote_bss_start[], __mote_bss_end[];                           \
-    extern const uint16_t mote_game_icon_data[] __attribute__((weak));         \
+    extern const uint8_t mote_game_icon_data[] __attribute__((weak));          \
     __attribute__((section(".mote_header"), used))                            \
     const MoteModuleHeader mote_module_header = {                                         \
         .magic = MOTE_MODULE_MAGIC,                                                \
